requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.0
plotly==5.17.0
Pillow==10.1.0
joblib==1.3.2